pyproject.toml
[project]
name = "tarea"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "app>=0.0.1",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "flet>=0.82.2",
    "mysql-connector-python>=9.6.0",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
]

[project.scripts]

#aqui definimos el comando corto
#quitamos el src porque definimos "src" como la raiz

app="main:main"

[tool.uv]
package = true

#Esta secion es crucial para que python vea el ontenido de src como el nivel raiz 
[tool.setuptools.package-dir]
"" = "src"
